Reorganize folder structure to match Vue idioms
