+++ assets/sounds/SS_beaconFound2_48k.wav
