+++ assets/sounds/beacons/Classic/Classic_OnAxis.wav
