+++ index.html
@@ -1,9 +1,9 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <link rel="apple-touch-icon" sizes="180x180" href="/src/icons/apple-touch-icon.png">
-    <link rel="icon" type="image/png" sizes="32x32" href="/src/icons/favicon-32x32.png">
-    <link rel="icon" type="image/png" sizes="16x16" href="/src/icons/favicon-16x16.png">
+    <link rel="apple-touch-icon" sizes="180x180" href="/assets/icons/apple-touch-icon.png">
+    <link rel="icon" type="image/png" sizes="32x32" href="/assets/icons/favicon-32x32.png">
+    <link rel="icon" type="image/png" sizes="16x16" href="/assets/icons/favicon-16x16.png">
 
     <meta charset="UTF-8" />
     <!-- Render to a usable size on mobile -->
@@ -11,10 +11,10 @@
 
     <link rel="stylesheet" href="https://unpkg.com/leaflet/dist/leaflet.css" />
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.4/css/all.min.css" />
-    <link rel="stylesheet" href="/src/css/main.css" />
+    <link rel="stylesheet" href="/src/main.css" />
 
-    <script src="/src/js/vendor/unmute.js" type="module"></script>
-    <script src="/src/js/main.js" type="module"></script>
+    <script src="/src/vendor/unmute.js" type="module"></script>
+    <script src="/src/main.js" type="module"></script>
   </head>
 
   <body></body>
